LPD-18312 Use separated files for utils
