+++ modules/test/playwright/helpers/HeadlessCommerceAdminChannelApiHelper.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {getRandomInt} from '../utils/util';
+import {getRandomInt} from '../utils/getRandomInt';
 import {ApiHelpers} from './ApiHelpers';
 
 type TChannel = {
