+++ modules/test/playwright/pages/export-import-web/exportImportFrame.page.ts
@@ -7,7 +7,7 @@
 
 import {Page} from '@playwright/test';
 
-import {zipFolder} from '../../utils/util';
+import {zipFolder} from '../../utils/zipFolder';
 
 export class ExportImportFramePage {
 	readonly page: Page;
