+++ modules/test/playwright/utils/getRandomInt.ts
@@ -1,0 +1,7 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+export function getRandomInt(): number {
+	return Math.floor(Math.random() * 9999999999);
+}
