+++ modules/test/playwright/utils/zipFolder.ts
@@ -7,10 +7,6 @@ import * as os from 'node:os'; // eslint-disable-line @liferay/no-extraneous-dep
 import * as path from 'path';
 import {zip} from 'zip-a-folder';
 
-export function getRandomInt(): number {
-	return Math.floor(Math.random() * 9999999999);
-}
-
 export async function zipFolder(folderPath: string) {
 	const tempFilePath = path.join(os.tmpdir(), path.basename(folderPath));
 	await zip(folderPath, tempFilePath);
